fix: criação e recarregamento de comentários

+++ server/src/modules/comment/comment.controller.ts
@@ -4,8 +4,13 @@ import { CreateCommentService, DeleteCommentService, EditCommentService, LikeCom
 import { ListCommentsByPostService } from "./services/CommentList.service.js";
 
 // Criar comentários
-export async function CreateCommentController(req: CustomRequest, res: Response){
-    try{
+export async function CreateCommentController(req: CustomRequest, res: Response) {
+    try {
+
+        if (!req.user) {
+            throw new Error("Usuário não encontrado");
+        }
+
         const author = req.user._id;
 
         const postId = req.params.id;
@@ -14,32 +19,42 @@ export async function CreateCommentController(req: CustomRequest, res: Response)
 
         const result = await CreateCommentService(author, content, postId);
 
-        res.status(201).json({msg: result.msg, comment: result.comment});
+        res.status(201).json({ msg: result.msg, comment: result.comment });
     }
-    catch(err: any){
-        res.status(500).json({error: err.message});
+    catch (err: any) {
+        res.status(500).json({ error: err.message });
     }
 }
 
 // Deletar comentário
 export async function DeleteCommentController(req: CustomRequest, res: Response) {
-    try{
+    try {
+
+        if (!req.user) {
+            throw new Error("Usuário não encontrado");
+        }
+
         const commentId = req.params.id;
 
         const authorId = req.user._id;
 
         const result = await DeleteCommentService(commentId, authorId);
 
-        res.status(200).json({msg: result.msg});
+        res.status(200).json({ msg: result.msg });
     }
-    catch(err: any){
-        res.status(500).json({error: err.message});
+    catch (err: any) {
+        res.status(500).json({ error: err.message });
     }
 }
 
 // Editar comentário
-export async function EditCommentController(req: CustomRequest, res: Response){
-    try{
+export async function EditCommentController(req: CustomRequest, res: Response) {
+    try {
+
+        if (!req.user) {
+            throw new Error("Usuário não encontrado");
+        }
+
         const authorId = req.user._id;
 
         const commentId = req.params.id;
@@ -48,32 +63,41 @@ export async function EditCommentController(req: CustomRequest, res: Response){
 
         const result = await EditCommentService(authorId, commentId, updates);
 
-        res.status(200).json({msg: result.msg, comment: result.comment});
+        res.status(200).json({ msg: result.msg, comment: result.comment });
     }
-    catch(err: any){
-        res.status(500).json({error: err.message});
+    catch (err: any) {
+        res.status(500).json({ error: err.message });
     }
 }
 
 // Dar like em comentário
-export async function LikeCommentController(req: CustomRequest, res: Response){
-    try{
+export async function LikeCommentController(req: CustomRequest, res: Response) {
+    try {
+
+        if (!req.user) {
+            throw new Error("Usuário não encontrado");
+        }
+
         const commentId = req.params.id;
 
         const userId = req.user._id;
 
         const result = await LikeCommentService(userId, commentId);
 
-        res.status(200).json({msg: result.msg, liked: result.liked});
+        res.status(200).json({ msg: result.msg, liked: result.liked });
     }
-    catch(err: any){
-        res.status(500).json({error: err.message});
+    catch (err: any) {
+        res.status(500).json({ error: err.message });
     }
 }
 
 // Listar comentários de uma postagem
-export async function ListCommentsByPostController(req: CustomRequest, res: Response){
-    try{
+export async function ListCommentsByPostController(req: CustomRequest, res: Response) {
+    try {
+
+        if (!req.user) {
+            throw new Error("Usuário não encontrado");
+        }
         const postId = req.params.id;
 
         const userId = req.user._id;
@@ -82,7 +106,7 @@ export async function ListCommentsByPostController(req: CustomRequest, res: Resp
 
         res.status(200).json({ msg: result.comments });
     }
-    catch(err: any){
-        res.status(500).json({error: err.message});
+    catch (err: any) {
+        res.status(500).json({ error: err.message });
     }
 }
